refactor(analysis): extract specialized modules for linting, testing, and type checking
Decompose analysis.py by extracting three specialized modules:

- linting: ruff tool integration for code quality checks
- testing: test execution and functional correctness verification
- type_checking: metrics computation for typing coverage and documentation

This improves modularity and separation of concerns, making the codebase more maintainable and testable.

Changes:
- Create src/linting module with ruff tool utilities
- Create src/testing module with test execution logic
- Create src/type_checking module with AST-based metrics
- Update analysis.py to use new modules via imports
- Register new packages in pyproject.toml for hatch build
- Remove ruff from project dependencies (now external tool)
