+++ pyproject.toml
@@ -37,7 +37,12 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-packages = ["src/robofactor"]
+packages = [
+  "src/robofactor",
+  "src/linting",
+  "src/type_checking",
+  "src/testing",
+]
 
 [dependency-groups]
 dev = [
