+++ src/linting/__init__.py
@@ -1,0 +1,7 @@
+from .ruff_tool import format_lint_issue, is_complexity_issue, run_ruff_json
+
+__all__ = [
+    "format_lint_issue",
+    "is_complexity_issue",
+    "run_ruff_json",
+]
