+++ src/testing/__init__.py
@@ -1,0 +1,3 @@
+from .execution import check_functional_correctness
+
+__all__ = ["check_functional_correctness"]
