+++ src/type_checking/__init__.py
@@ -1,0 +1,3 @@
+from .metrics import docstring_and_typing_scores
+
+__all__ = ["docstring_and_typing_scores"]
